_archive/jarvis-safe/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
groq==0.9.0
websockets==13.0
pydantic==2.9.0
